Fix renew_certificate.sh script execution in crontab.
While the `renew_certificate.sh` script worked fine when I ran it manually, it
didn't appear to be running via the `crontab` (based on the lack of StatsD
metrics going to my DataDog agent). After piping the output of the `crontab` to
a file, I found two issues. First, the output was:

> the input device is not a TTY

This was caused by the `-it` options in my `docker run` command. The `-t`
requires a TTY interface to be allocated; which, is not allocated when running
via `crontab`.

Once I removed the `-it` options, I was faced with a second error in the log
output:

```
Traceback (most recent call last):
  File "bin/docker-compose", line 6, in <module>
  File "compose/cli/main.py", line 71, in main
  File "compose/cli/main.py", line 124, in perform_command
  File "compose/cli/main.py", line 467, in exec_command
  File "site-packages/dockerpty/pty.py", line 338, in start
  File "site-packages/dockerpty/io.py", line 32, in set_blocking
ValueError: file descriptor cannot be a negative integer (-1)
Failed to execute script docker-compose
```

Apparently, this is a known issue with `docker-compose` being run inside a
script, like `crontab`. To work around this, adding the `-T` option tells
docker not to allocate a Pseudo-TTY for `docker-compose`.

Source: https://stackoverflow.com/questions/43099116/the-input-device-is-not-a-tty
Source: https://github.com/docker/compose/issues/3352